we need more lints! MOAAAAAR~

+++ src/scsi/freebsd.rs
@@ -10,11 +10,11 @@ use scsi::SCSIDevice;
 use std::io::Error;
 
 impl SCSIDevice for Device {
-	fn do_cmd(&self, cmd: &[u8], dir: Direction, sense_len: u8, data_len: usize)-> Result<(Vec<u8>, Vec<u8>), Error> {
+	fn do_cmd(&self, cmd: &[u8], dir: Direction, sense_len: usize, data_len: usize)-> Result<(Vec<u8>, Vec<u8>), Error> {
 		// might've used Vec::with_capacity(), but this requires rebuilding with Vec::from_raw_parts() later on to hint actual size of data in buffer vecs,
 		// and we're not expecting this function to be someone's bottleneck
-		let mut sense = vec![0; sense_len as usize];
-		let mut data = vec![0; data_len as usize];
+		let mut sense = vec![0; sense_len];
+		let mut data = vec![0; data_len];
 
 		let timeout = 10; // in seconds; TODO configurable
 
@@ -45,6 +45,7 @@ impl SCSIDevice for Device {
 			csio.sense_len = sense.capacity() as u8;
 			csio.tag_action = MSG_SIMPLE_Q_TAG as u8;
 
+			#[allow(trivial_casts)] // XXX
 			libc::memcpy(
 				&mut csio.cdb_io.cdb_bytes as *mut _ as *mut c_void,
 				cmd.as_ptr() as *const c_void,
@@ -67,6 +68,7 @@ impl SCSIDevice for Device {
 				unsafe {
 					let csio = ccb.csio();
 
+					#[allow(trivial_casts)] // XXX
 					libc::memcpy(
 						sense.as_mut_ptr() as *mut c_void,
 						&csio.sense_data as *const _ as *const c_void,
